requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
networkx==3.2.1
celery==5.3.6
redis==5.0.1
python-multipart
python-jose[cryptography]
passlib
cryptography
python-dotenv
slowapi
